refactor(analysis): promote xbgn/xend to NMTool base class
Move the xbgn/xend x-axis window properties from NMToolSpike, NMToolEvent,
and NMToolFit into the NMTool base class. Each subclass now initialises
_xbgn/_xend from its config and references the inherited protected attrs
instead of name-mangled private ones. The duplicate property/setter code
(~100 lines) is removed from all three subclasses.

Consolidate the generic default and validation tests into a new
TestNMToolXWindow class in test_nm_tool.py; behaviour-specific tests
(detection/fit window restriction) remain in the subclass test files.
