+++ pyneuromatic/analysis/nm_tool_spike.py
@@ -113,10 +113,11 @@ class NMToolSpike(NMTool):
         self._results_to_cache = self._config.results_to_cache
         self._results_to_numpy = self._config.results_to_numpy
 
+        self._xbgn = self._config.xbgn
+        self._xend = self._config.xend
+
         self.__ylevel: float = 0.0
         self.__func_name: str = "level+"
-        self.__xbgn: float = -math.inf
-        self.__xend: float = math.inf
 
         # Internal run state — reset by run_init()
         self._spike_times: list[np.ndarray] = []
@@ -176,42 +177,6 @@ class NMToolSpike(NMTool):
         nmh.history("set func_name=%r" % self.__func_name, quiet=quiet)
         nmch.add_nm_command("%s.func_name = %r" % (self._name, self.__func_name))
 
-    @property
-    def xbgn(self) -> float:
-        """X-axis window start for detection. Default ``-inf`` (no lower bound)."""
-        return self.__xbgn
-
-    @xbgn.setter
-    def xbgn(self, value: float) -> None:
-        self._xbgn_set(value)
-
-    def _xbgn_set(self, value: float, quiet: bool = nmc.QUIET) -> None:
-        if isinstance(value, bool) or not isinstance(value, (int, float)):
-            raise TypeError(nmu.type_error_str(value, "xbgn", "float"))
-        if math.isnan(value):
-            raise ValueError("xbgn cannot be NaN")
-        self.__xbgn = float(value)
-        nmh.history("set xbgn=%g" % self.__xbgn, quiet=quiet)
-        nmch.add_nm_command("%s.xbgn = %r" % (self._name, self.__xbgn))
-
-    @property
-    def xend(self) -> float:
-        """X-axis window end for detection. Default ``+inf`` (no upper bound)."""
-        return self.__xend
-
-    @xend.setter
-    def xend(self, value: float) -> None:
-        self._xend_set(value)
-
-    def _xend_set(self, value: float, quiet: bool = nmc.QUIET) -> None:
-        if isinstance(value, bool) or not isinstance(value, (int, float)):
-            raise TypeError(nmu.type_error_str(value, "xend", "float"))
-        if math.isnan(value):
-            raise ValueError("xend cannot be NaN")
-        self.__xend = float(value)
-        nmh.history("set xend=%g" % self.__xend, quiet=quiet)
-        nmch.add_nm_command("%s.xend = %r" % (self._name, self.__xend))
-
     # ------------------------------------------------------------------
     # Lifecycle
 
@@ -244,8 +209,8 @@ class NMToolSpike(NMTool):
         _indexes, x_times = find_level_crossings_nmdata(
             data, self.__ylevel,
             func_name=self.__func_name,
-            xbgn=self.__xbgn,
-            xend=self.__xend,
+            xbgn=self._xbgn,
+            xend=self._xend,
             ignore_nans=self._ignore_nans,
         )
         self._spike_times.append(x_times)
@@ -308,7 +273,7 @@ class NMToolSpike(NMTool):
                 d,
                 "NMSpike(source=%s, ylevel=%g, func_name=%r, xbgn=%s, xend=%s, n=%d)"
                 % (name, self.__ylevel, self.__func_name,
-                   self.__xbgn, self.__xend, len(times)),
+                   self._xbgn, self._xend, len(times)),
             )
         counts = np.array([len(t) for t in self._spike_times], dtype=float)
         d_count = f.data.new("SP_count", nparray=counts)
@@ -316,7 +281,7 @@ class NMToolSpike(NMTool):
             d_count,
             "NMSpike(ylevel=%g, func_name=%r, xbgn=%s, xend=%s, n_epochs=%d)"
             % (self.__ylevel, self.__func_name,
-               self.__xbgn, self.__xend, len(self._epoch_names)),
+               self._xbgn, self._xend, len(self._epoch_names)),
         )
         f.data.new(
             "SP_epoch_names",
